+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldBashEnchantment.java
@@ -2,7 +2,7 @@ package com.insane96mcp.shieldsplus.world.item.enchantment;
 
 import com.insane96mcp.shieldsplus.setup.SPEnchantments;
 import com.insane96mcp.shieldsplus.setup.Strings;
-import insane96mcp.insanelib.util.MCUtils;
+import com.insane96mcp.shieldsplus.world.item.SPShieldItem;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.sounds.SoundEvents;
@@ -10,26 +10,18 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.phys.AABB;
-import net.minecraftforge.common.ForgeMod;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
-import java.util.UUID;
 
 public class ShieldBashEnchantment extends Enchantment {
 
-	static UUID STEP_HEIGHT_ON_BASH = UUID.fromString("259f26e7-6914-4856-a01f-8b2295e6d244");
-
 	public ShieldBashEnchantment() {
-		super(Rarity.VERY_RARE, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
+		super(Rarity.VERY_RARE, SPShieldItem.SHIELD, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
 
 	public int getMinCost(int level) {
@@ -47,11 +39,6 @@ public class ShieldBashEnchantment extends Enchantment {
 	@Override
 	public boolean isTreasureOnly() {
 		return true;
-	}
-
-	@Override
-	public boolean canEnchant(ItemStack stack) {
-		return stack.getItem() instanceof ShieldItem;
 	}
 
 	@Override
@@ -75,7 +62,6 @@ public class ShieldBashEnchantment extends Enchantment {
 			}
 			else if (player.isCrouching() && player.isOnGround()) {
 				tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) 12);
-				MCUtils.applyModifier(player, ForgeMod.STEP_HEIGHT_ADDITION.get(), STEP_HEIGHT_ON_BASH, "Shield Bash step height", 0.6d, AttributeModifier.Operation.ADDITION, false);
 				float f = -Mth.sin(player.getYRot() * ((float) Math.PI / 180F));
 				float f1 = Mth.cos(player.getYRot() * ((float) Math.PI / 180F));
 				player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 1.3f);
@@ -92,11 +78,6 @@ public class ShieldBashEnchantment extends Enchantment {
 
 		if (tag.getByte(Strings.Tags.SHIELD_BASHING) > 0) {
 			tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) (tag.getByte(Strings.Tags.SHIELD_BASHING) - 1));
-			if (tag.getByte(Strings.Tags.SHIELD_BASHING) == 0) {
-				if (player.getAttribute(ForgeMod.STEP_HEIGHT_ADDITION.get()) != null)
-					//noinspection ConstantConditions
-					player.getAttribute(ForgeMod.STEP_HEIGHT_ADDITION.get()).removeModifier(UUID.fromString("259f26e7-6914-4856-a01f-8b2295e6d244"));
-			}
 			if (player.isBlocking()) {
 				damageAndKnockback(player);
 			}